+++ R/residuals.R
@@ -44,5 +44,7 @@ Defaulting to `type="response"`', type, model_sum(object)))
   .resid <- split(.resid, col(.resid))
   nm <- if(length(.resid) == 1) ".resid" else map_chr(object$response, expr_name)
   
-  transmute(object$data, !!!set_names(.resid, nm))
+  out <- object$data[index_var(object$data)]
+  out[nm] <- .resid
+  out
 }