+++ NEWS.md
@@ -1,3 +1,5 @@
+# fabletools (development version)
+
 # fabletools 0.2.0
 
 ## New features