+++ NEWS.md
@@ -33,6 +33,7 @@
 * Fixed issue with aggregated date vectors (#230).
 * Fixed display of models in `View()` panel.
 * Fixed issue with combination models not inheriting vctrs functionality (#237).
+* `aggregate_key()` can now be used with non-syntactic variable names.
 
 # fabletools 0.2.0