+++ NEWS.md
@@ -1,4 +1,4 @@
-# fabletools (development version)
+# fabletools 0.2.1
 
 ## New features